- Split event management into 2 fragments - Fixed some navigation issues

+++ app/src/main/java/spontaniius/ui/find_event/FindEventFragment.kt
@@ -302,7 +302,7 @@ class FindEventFragment : Fragment() {
                         eventView.details.setOnClickListener {
                             val action =
                                 FindEventFragmentDirections.actionFindEventFragmentToEventManagementFragment(
-                                    eventId = eventView.eventId,
+                                    eventId = eventView.eventId.toInt(),
                                     isEventOwner = false
                                 )
                             findNavController().navigate(action)
